hotel.js, hotel.css and assets folder created
