+++ src/App.js
@@ -1,10 +1,16 @@
 import logo from './logo.svg';
 import './App.css';
+import React from 'react';
+
 
 
 function App() {
   return (
     <div className="App">
+      <hotels/>
+      <h1>Travel</h1>
+      <image src = {logo} className = "App-logo" alt = "logo"/>
+      
       
     </div>
   );
